+++ R/spct.fshift.r
@@ -9,6 +9,11 @@
 #' zero shift (= additive correction) to the values in the returned object.
 #' Metadata attributes are retained unchanged.
 #'
+#' @note Method \code{fshift} is not implemented for \code{solute_spct} objects
+#'   as the spectral data stored in them are a description of an intensive
+#'   property of a substance. To represent solutions of specific concentrations
+#'   of solutes, \code{filter_spct} objects can be used.
+#'
 #' @param x An R object
 #' @param ... additional named arguments passed down to \code{f}.
 #'